# perf: e2 mul on koalabear