--- a/field/generator/internal/templates/extensions/e2.go.tmpl
+++ b/field/generator/internal/templates/extensions/e2.go.tmpl
@@ -310,7 +310,8 @@ func (z *E2) Mul(x, y *E2) *E2 {
     {{- if eq .FF "koalabear"}}
 		d.Add(&d, &c)
 		a.Sub(&a, &d)
-		d.Add(&d, &c).Add(&d, &c)
+        b.Double(&c)
+        d.Add(&d, &b)
 		z.A0 = d
 		z.A1 = a
     {{- else if eq .FF "babybear"}}