# Description

tiny change in e2 mul; karatsuba already computes `a+b` so we use `Double` to compute `2b` and add it to the sum to have `a+3b` where `3` is the quadratic non-residue. This yields -10%.

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?

Current tests pass.

# How has this been benchmarked?

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkE2Mul-8              16.1          14.4          -10.52%
BenchmarkE4Mul-8              27.6          26.3          -4.71%
```

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules