@gbotrel 2025-09-15
on my M1 Max:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkE4Mul-10     22.5          22.5          +0.09%
BenchmarkE4Mul-10     22.5          22.5          -0.13%
BenchmarkE4Mul-10     22.5          22.5          -0.18%
BenchmarkE4Mul-10     22.5          22.5          -0.31%
```